requirements.txt
Django==5.1.7
djangorestframework==3.15.2
drf-yasg==1.21.10
djangorestframework-simplejwt==5.3.1